+++ gems.rb
@@ -9,15 +9,19 @@ gemspec
 
 group :maintenance, optional: true do
 	gem "bake-gem"
+	gem "bake-releases"
+	
+	gem "agent-context"
 	
 	gem "utopia-project"
-	gem "bake-releases"
 end
 
 group :test do
 	gem "sus"
 	gem "covered"
 	gem "decode"
+	
+	gem "sus-fixtures-async"
 	
 	gem "rubocop"
 	gem "rubocop-md"