+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Make `modernize:releases` part of the default `modernize` task.
+
 ## v0.52.0
 
   - Add support for automatic GitHub release generation.